tracker/requirements.txt
requests
pyserial
python-dotenv
haversine
bleak
aio-pika